# docs/api/worksession-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Ergo.WorkSession.API",
    "version": "1.0"
  },
  "paths": {
    "/api/work-session/postures/{userId}": {
      "get": {
        "tags": [
          "WorkSession"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/work-session/postures": {
      "post": {
        "tags": [
          "WorkSession"
        ],
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostureRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/work-session/postures/{postureId}": {
      "delete": {
        "tags": [
          "WorkSession"
        ],
        "parameters": [
          {
            "name": "postureId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/work-session/calibration/calculate": {
      "post": {
        "tags": [
          "WorkSession"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/work-session/session/start": {
      "post": {
        "tags": [
          "WorkSession"
        ],
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/work-session/session/{sessionId}/monitor": {
      "post": {
        "tags": [
          "WorkSession"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/work-session/session/{sessionId}/pause": {
      "post": {
        "tags": [
          "WorkSession"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/work-session/session/{sessionId}/resume": {
      "post": {
        "tags": [
          "WorkSession"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/work-session/session/{sessionId}/stop": {
      "post": {
        "tags": [
          "WorkSession"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreatePostureRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "alias": {
            "type": "string",
            "nullable": true
          },
          "vector": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "double"
            },
            "nullable": true
          },
          "isPersistent": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "SessionMode": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "StartSessionRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "mode": {
            "$ref": "#/components/schemas/SessionMode"
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "postureId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "temporaryVector": {
            "type": "array",
            "items": {
              "type": "number",
              "format": "double"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}